+++ plugins/web-interface/static/js/state.js
@@ -1,0 +1,21 @@
+/**
+ * Manages the global state for the web-interface plugin of Pup.
+ *
+ * @file static/js/state.js
+ */
+
+class ProcessSelector {
+  constructor() {
+    this.selectedProcessId = null
+  }
+  set(id) {
+    this.selectedProcessId = id
+  }
+  get() {
+    return this.selectedProcessId
+  }
+}
+
+export const processStatusInventory = new Map()
+export const processConfigInventory = new Map()
+export const processSelector = new ProcessSelector()