+++ plugins/web-interface/static/js/ui.js
@@ -5,12 +5,37 @@
  * @file static/js/ui.js
  */
 
-import { ansiToHtml, getStatusColor, ProcessStateToString, showSpecificClassElements } from "./helpers.js"
+import { ansiToHtml, formatBytes, formatTime, getStatusColor, ProcessStateToString, showSpecificClassElements } from "./helpers.js"
+import { processConfigInventory, processSelector, processStatusInventory } from "./state.js"
 
-let selectedProcessId
-
-const processStatusInventory = new Map()
-const processConfigInventory = new Map()
+async function updateInstance(instanceData, processData) {
+  const systemMemoryUsagePercent = ((instanceData.systemMemory.total - instanceData.systemMemory.free) / instanceData.systemMemory.total * 100).toFixed(2)
+  setTextContentBySelector("#system-memory-usage-percent", `${systemMemoryUsagePercent}%`)
+  if (systemMemoryUsagePercent > 90) {
+    document.getElementById("system-memory-container").className = "bg-bad fg-bad"
+  } else if (systemMemoryUsagePercent > 80) {
+    document.getElementById("system-memory-container").className = "bg-warning fg-warning"
+  }
+  const systemMemoryUsage = instanceData.systemMemory.total - instanceData.systemMemory.free
+  setTextContentBySelector("#system-memory-usage", `${formatBytes(systemMemoryUsage, false, "GB")} / ${formatBytes(instanceData.systemMemory.total, true, "GB")}`)
+  const systemSwapUsagePercent = ((instanceData.systemMemory.swapTotal - instanceData.systemMemory.swapFree) / instanceData.systemMemory.swapTotal * 100).toFixed(2)
+  setTextContentBySelector("#system-swap-usage-percent", `${systemSwapUsagePercent}%`)
+  const systemSwapUsage = instanceData.systemMemory.swapTotal - instanceData.systemMemory.swapFree
+  setTextContentBySelector("#system-swap-usage", `${formatBytes(systemSwapUsage, false, "GB")} / ${formatBytes(instanceData.systemMemory.swapTotal, true, "GB")}`)
+  setTextContentBySelector("#system-uptime", formatTime(instanceData.osUptime))
+  setTextContentBySelector("#system-os-release", instanceData.osRelease)
+  setTextContentBySelector("#system-load", instanceData.loadAvg.map((l) => l.toFixed(2)).join(", "))
+  setTextContentBySelector("#pup-memory-usage", formatBytes(instanceData.memory.rss, true))
+  setTextContentBySelector("#pup-deno-version", instanceData.denoVersion.deno)
+  setTextContentBySelector("#pup-version", instanceData.version)
+  const uptime = new Date() - new Date(instanceData.started)
+  setTextContentBySelector("#pup-uptime", formatTime(uptime / 1000))
+  setTextContentBySelector("#pup-processes", instanceData.processes.length.toString())
+  processData.forEach((currentProcess) => {
+    updateProcessCard(currentProcess)
+  })
+  changeLogScope(processSelector.get())
+}
 
 /**
  * Change the log scope to a specific process.
@@ -19,45 +44,54 @@ const processConfigInventory = new Map()
  * @param {string} processId - The id of the process.
  */
 export function changeLogScope(processId) {
-  selectedProcessId = processId
+  processSelector.set(processId)
 
-  // Remove the class from all process cards
-  const processCards = document.querySelectorAll(".process-card")
-  processCards.forEach((processCard) => {
-    processCard.classList.remove("process-card-selected")
-  })
+  let sanitizedId
 
-  // Add the class to the selected process card
-  const sanitizedId = sanitizeId(processId)
-  const selectedProcessCard = document.getElementById(`process-card-${sanitizedId}`)
-  if (selectedProcessCard) selectedProcessCard.classList.add("process-card-selected")
+  if (processId) {
+    sanitizedId = sanitizeId(processId)
+
+    // Show actions
+    document.getElementById("actions-container").classList.remove("hidden")
+
+    // Remove the class from all process cards
+    const processCards = document.querySelectorAll(".process-card")
+    processCards.forEach((processCard) => {
+      processCard.classList.remove("process-card-selected")
+    })
+
+    // Add the class to the selected process card
+    const selectedProcessCard = document.getElementById(`process-card-${sanitizedId}`)
+    if (selectedProcessCard) selectedProcessCard.classList.add("process-card-selected")
+
+    setTextContentBySelector("#toolbar-header", processId)
+
+    // Update the toolbar with the selected process details
+    updateSidebar(sanitizedId)
+  } else {
+    setTextContentBySelector("#toolbar-header", "Global Pup logs")
+
+    // Hide actions
+    document.getElementById("actions-container").classList.add("hidden")
+
+    updateSidebar()
+  }
 
   // Show logs for the selected process only
   showSpecificClassElements("logs", sanitizedId)
-
-  // Update the text of #selected-process-id
-  const selectedProcessIdElement = document.getElementById("selected-process-id")
-  if (selectedProcessIdElement) {
-    selectedProcessIdElement.textContent = processId
-  } else {
-    console.warn("Cannot find #selected-process-id element to update text.")
-  }
-
-  // Update the toolbar with the selected process details
-  updateToolbar(sanitizedId)
 }
 
 /**
  * Add a log entry to the logs section.
  *
  * @param {Object} log - Log entry.
- * @param {string} selectedProcessId - The id of the selected process.
+ * @param {string} processId - The id of the selected process.
  */
-function addLog(log, selectedProcessId) {
+function addLog(log, processId) {
   const logsDiv = document.getElementById("logs")
   logsDiv.classList.remove("hidden")
 
-  const hiddenClass = selectedProcessId === log.id ? "" : "hidden"
+  const hiddenClass = processId === log.id ? "" : "hidden"
   let logsHtml = `<pre class="${log.id} ${hiddenClass}">${log.ts} | ${log.category} | ${log.severity} > ${ansiToHtml(log.text)}\n</pre>`
   logsDiv.innerHTML = logsHtml + logsDiv.innerHTML
 }
@@ -92,28 +126,36 @@ function sanitizeId(id) {
 }
 
 /**
- * Update the toolbar with the details of a specific process.
+ * Update the sidebar with the details of a specific process.
  *
  * @param {string} processId - The id of the process.
  */
-function updateToolbar(processId) {
-  // Get process config and status data
-  const processConfig = processConfigInventory.get(processId)
-  const processStatus = processStatusInventory.get(processId)
+function updateSidebar(processId) {
+  if (processId) {
+    // Get process config and status data
+    const processConfig = processConfigInventory.get(processId)
+    const processStatus = processStatusInventory.get(processId)
 
-  // Update process config values
-  setTextContentBySelector("#process-config-command", processConfig.cmd || processConfig.worker)
-  setTextContentBySelector("#process-config-autostart", processConfig.autostart)
-  setTextContentBySelector("#process-config-cron", processConfig.cron)
-  setTextContentBySelector("#process-config-restart", processConfig.restart || "always")
-  setTextContentBySelector("#process-config-terminate", processConfig.terminate)
+    // Update process config values
+    setTextContentBySelector("#process-config-command", processConfig.cmd || processConfig.worker)
+    setTextContentBySelector("#process-config-autostart", processConfig.autostart)
+    setTextContentBySelector("#process-config-cron", processConfig.cron)
+    setTextContentBySelector("#process-config-restart", processConfig.restart || "always")
+    setTextContentBySelector("#process-config-terminate", processConfig.terminate)
 
-  // Update process status values
-  setTextContentBySelector("#process-status-type", processStatus.type)
-  setTextContentBySelector("#process-status-blocked", processStatus.blocked)
-  setTextContentBySelector("#process-status-status", ProcessStateToString(processStatus.status))
-  setTextContentBySelector("#process-status-started", processStatus.started)
-  setTextContentBySelector("#process-status-updated", processStatus.updated)
+    // Update process status values
+    setTextContentBySelector("#process-status-type", processStatus.type)
+    setTextContentBySelector("#process-status-blocked", processStatus.blocked)
+    setTextContentBySelector("#process-status-status", ProcessStateToString(processStatus.status))
+    setTextContentBySelector("#process-status-started", processStatus.started)
+    setTextContentBySelector("#process-status-updated", processStatus.updated)
+
+    document.getElementById("aside-overview").className = "hidden"
+    document.getElementById("aside-process").className = ""
+  } else {
+    document.getElementById("aside-overview").className = ""
+    document.getElementById("aside-process").className = "hidden"
+  }
 }
 
 /**
@@ -124,7 +166,6 @@ function updateToolbar(processId) {
  */
 function updateProcessCard(processData) {
   const { status /*, config */ } = processData
-
   const sanitizedId = sanitizeId(status.id)
 
   // Get or create process card
@@ -136,6 +177,13 @@ function updateProcessCard(processData) {
       changeLogScope(sanitizedId)
     })
 
+    // Add icon
+    const icon = document.createElement("i")
+    icon.classList.add("fas")
+    if (status.type === "process") icon.classList.add("fa-cube")
+    else if (status.type === "cluster") icon.classList.add("fa-cubes")
+    processCard.appendChild(icon)
+
     // Add process title
     const title = document.createElement("span")
     title.classList.add("process-title")
@@ -144,6 +192,8 @@ function updateProcessCard(processData) {
     // Add process status
     const ProcessState = document.createElement("span")
     ProcessState.classList.add("process-status")
+    ProcessState.classList.add("badge")
+
     title.appendChild(ProcessState)
 
     processCard.appendChild(title)
@@ -159,12 +209,15 @@ function updateProcessCard(processData) {
 
   // Update process status
   const ProcessState = processCard.querySelector(".process-status")
-  ProcessState.textContent = ` - ${ProcessStateToString(status.status)}`
+  ProcessState.textContent = `${ProcessStateToString(status.status)}`
+  ProcessState.className = "process-status badge"
+  ProcessState.classList.add(`bg-${getStatusColor(status.status)}`)
+  ProcessState.classList.add(`border-${getStatusColor(status.status)}`)
 
   // Set border color depending on process status
-  processCard.className = `process-card border-${getStatusColor(status.status)}`
+  processCard.className = `process-card`
 
   return processCard
 }
 
-export { addLog, processConfigInventory, processStatusInventory, selectedProcessId, updateProcessCard, updateToolbar }
+export { addLog, updateInstance, updateProcessCard, updateSidebar }