+++ docs/changelog.md
@@ -16,8 +16,9 @@ All notable changes to this project will be documented in this section.
 - chore(build): Add build step for bundling web interface assets.
 - feature(webinterface): Add "toolbar" showing config/status details of selected process.
 - feature(webinterface): Give process "tabs" colors visualising process status.
-- chore(webinterface): Code cleanup, convert js to esm etc.
+- chore(webinterface): Code refactor, convert js to esm etc.
 - chore(webinterface): Add separate README.md (plugins/web-interface/README.md) with docs for developers/contributors.
+- feature(status,logger): Use Deno KV instead of plain file, for keeping status and internal logs.
 
 ## [1.0.0-beta.32] - 2023-06-06